+++ src/form/b-checkbox/CHANGELOG.md
@@ -11,6 +11,10 @@ Changelog
 
 ## v3.0.0-rc.91 ()
 
+#### :boom: Breaking Change
+
+* Now `toggle` returns a component value
+
 #### :house: Internal
 
 * Fixed ESLint warnings